+++ src/Make_morph.mak
@@ -69,6 +69,7 @@ SRC =	arabic.c						\
 	screen.c						\
 	search.c						\
 	sha256.c						\
+	sign.c							\
 	spell.c							\
 	spellfile.c						\
 	syntax.c						\
