+++ src/proto/buffer.pro
@@ -69,23 +69,6 @@ char_u *buf_spname(buf_T *buf);
 void switch_to_win_for_buf(buf_T *buf, win_T **save_curwinp, tabpage_T **save_curtabp, bufref_T *save_curbuf);
 void restore_win_for_buf(win_T *save_curwin, tabpage_T *save_curtab, bufref_T *save_curbuf);
 int find_win_for_buf(buf_T *buf, win_T **wp, tabpage_T **tp);
-void init_signs(void);
-int sign_group_get_next_signid(buf_T *buf, char_u *groupname);
-int sign_in_group(signlist_T *sign, char_u *group);
-dict_T *sign_get_info(signlist_T *sign);
-void buf_addsign(buf_T *buf, int id, char_u *groupname, int prio, linenr_T lnum, int typenr);
-linenr_T buf_change_sign_type(buf_T *buf, int markId, char_u *group, int typenr);
-int buf_getsigntype(buf_T *buf, linenr_T lnum, int type);
-linenr_T buf_delsign(buf_T *buf, linenr_T atlnum, int id, char_u *group);
-int buf_findsign(buf_T *buf, int id, char_u *group);
-signlist_T *buf_getsign_with_id(buf_T *buf, int id, char_u *group);
-int buf_findsign_id(buf_T *buf, linenr_T lnum, char_u *groupname);
-int buf_findsigntype_id(buf_T *buf, linenr_T lnum, int typenr);
-int buf_signcount(buf_T *buf, linenr_T lnum);
-void buf_delete_signs(buf_T *buf, char_u *group);
-void buf_delete_all_signs(char_u *groupname);
-void sign_list_placed(buf_T *rbuf, char_u *sign_group);
-void sign_mark_adjust(linenr_T line1, linenr_T line2, long amount, long amount_after);
 void set_buflisted(int on);
 int buf_contents_changed(buf_T *buf);
 void wipe_buffer(buf_T *buf, int aucmd);
